refactor: potraits
