+++ apps/web/components/ContactPage/ContactPage.tsx
@@ -1,8 +1,7 @@
 'use client'
 
-import { useId } from 'react'
+import { useState } from 'react'
 
-import { Tooth } from '@sdlgr/assets'
 import { Link } from '@sdlgr/link'
 import { useIsBot } from '@sdlgr/use-is-bot'
 
@@ -10,62 +9,33 @@ import { NoSSR } from '@web/components/NoSSR/NoSSR'
 import { useClientTranslation } from '@web/i18n/client'
 
 import {
-  PortraitContainer,
-  RotateTooth,
   StyledBody,
   StyledContactInfo,
   StyledContactInfoWrapper,
   StyledHeading,
   StyledLabel,
   StyledMainPortal,
-  StyledPortrait,
-  StyledPortraitContainer,
-  StyledTooth,
-  ToothHoleImage,
 } from './ContactPage.styles'
+import { Portrait } from './components/Portrait/Portrait'
+import { ToothLessPortrait } from './components/ToothLessPortrait/ToothLessPortrait'
 
 export function ContactPage() {
   const { t } = useClientTranslation('contactPage')
-  const id = useId()
+  const [painMode, setPainMode] = useState(false)
 
   const { isBot, isLoading } = useIsBot()
 
   const mail = 'sauldeleonguerrero@gmail.com'
 
-  const toothImages = [
-    '/assets/toothPlace-1.png',
-    '/assets/toothPlace-2.png',
-    '/assets/toothPlace-3.png',
-    '/assets/toothPlace-4.png',
-    '/assets/toothPlace-5.png',
-    '/assets/toothPlace-6.png',
-    '/assets/toothPlace-7.png',
-    '/assets/toothPlace-8.png',
-  ]
-
   return (
     <StyledMainPortal>
       <StyledContactInfoWrapper>
         <StyledContactInfo $isVisible={!isLoading}>
-          <PortraitContainer>
-            {toothImages.map((_, index) => (
-              <StyledTooth key={`${id}-${index}`} $index={index}>
-                <RotateTooth>
-                  <Tooth width={10} height={10} />
-                </RotateTooth>
-              </StyledTooth>
-            ))}
-
-            <StyledPortraitContainer>
-              <StyledPortrait
-                src="/assets/portrait.jpg"
-                alt={t('profilePicture')}
-                width={140}
-                height={140}
-              />
-              <ToothHoleImage $images={toothImages} />
-            </StyledPortraitContainer>
-          </PortraitContainer>
+          {painMode ? (
+            <ToothLessPortrait onClick={() => setPainMode(false)} />
+          ) : (
+            <Portrait onClick={() => setPainMode(true)} />
+          )}
           <StyledHeading $level={2}>Software Engineer</StyledHeading>
           <Link
             href="https://www.linkedin.com/in/sauldeleonguerrero"
