+++ apps/web/i18n/locales/en/contactPage.json
@@ -1,5 +1,6 @@
 {
   "profilePicture": "My profile picture",
+  "toothlessProfilePicture": "My toothless profile picture",
   "phoneAria": "My phone number",
   "phone": "Phone",
   "mail": "E-mail",
