+++ apps/web/i18n/locales/es/contactPage.json
@@ -1,5 +1,6 @@
 {
   "profilePicture": "Mi foto de perfil",
+  "toothlessProfilePicture": "My desdentada foto de perfil",
   "phoneAria": "Mi número de teléfono",
   "phone": "Tel",
   "mail": "E-mail",
